// src/constants/Enums.ts
export enum FlykraHeaders {
	ContentTypeJSON = 'application/json',
	UserAgent = 'FlykraClient/1.0.0',
	AcceptEncoding = 'gzip, deflate, br'
}

export enum FlykraBase {
	APIURL = 'https://flykra.pengubot.com'
}

export enum FetchTypes {
	GET = 'GET',
	POST = 'POST'
}

export enum GeneratorEndpoints {
	BATSLAP = 'batslap',
	BEAUTIFUL = 'beautiful',
	BOBROSS = 'bobross',
	EVIL = 'evil',
	FACEPALM = 'facepalm',
	GARBAGE = 'garbage',
	LIO = 'lio',
	MISSING = 'missing',
	RELIGION = 'religion',
	RIP = 'rip',
	TINDER = 'tinder',
	TRIGGERED = 'triggered',
	WANTED = 'wanted'
}

export enum OverlayEndpoints {
	RAINBOW = 'rainbow',
	APPROVED = 'sign/approved',
	REJECTED = 'sign/rejected'
}

export enum TextEndpoints {
	ACHIEVEMENT = 'achievement',
	CMM = 'cmm',
	ILLEGAL = 'illegal'
}